+++ wsi_service_base_plugins/openslide/wsi_service_plugin_openslide/slide.py
@@ -171,7 +171,7 @@ class Slide(BaseSlide):
                 tile_extent=self.__get_tile_extent(),
                 num_levels=len(levels),
                 levels=levels,
-                format=self.format.upper(),
+                format=self.format,
             )
             return slide_info
         except Exception as e: