+++ wsi_service_base_plugins/tiffslide/wsi_service_plugin_tiffslide/slide.py
@@ -147,7 +147,7 @@ class Slide(BaseSlide):
                 tile_extent=self.__get_tile_extent(),
                 num_levels=len(levels),
                 levels=levels,
-                format=self.format.upper(),
+                format=self.format,
             )
             return slide_info
         except HTTPException as e: