+++ wsi_service_base_plugins/wsidicom/wsi_service_plugin_wsidicom/__init__.py
@@ -1,6 +1,11 @@
+import glob
+import os
+
 from wsi_service_plugin_wsidicom.slide import Slide
 
-supported_file_extensions = ["dicom-folder"]
+
+def is_supported(filepath):
+    return len(glob.glob(os.path.join(filepath, "*.dcm"))) > 0
 
 
 async def open(filepath):