+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.11.0
+
+- New plugin concept
+- Extended slide info format content
+
 ## 0.10.20
 
 - Updated Dependencies